Change usize to u8 where sensible. Move controller input beginnings into new file. Add bundles for window and origin. Clean up systems file. Change XrController and XrTrackedObject markers

+++ src/head.rs
@@ -37,4 +37,4 @@ pub use crate::handedness::RightHanded;
 ///
 /// This component should be spawned including a  [`Camera3dBundle`] or similar and should be marked by a [`XrHandedness`] and one of the [`XrLeft`] or [`XrRight`] components.
 #[derive(Component)]
-pub struct XrEye(pub usize);
+pub struct XrEye(pub u8);
